complete makeupdate

+++ src/module/makeUpdate.py
@@ -44,7 +44,8 @@ def copy(srcPath, dstPath):
     dstDir = os.path.dirname(dstPath)
     if not os.path.exists(dstDir):
         os.makedirs(dstDir)
-    os.remove(dstPath)
+    if os.path.isfile(dstPath) and os.path.exists(dstPath):
+        os.remove(dstPath)
     dstFile = open(dstPath, "wb")
     temp = srcFile.read()
     srcFile.close()
@@ -61,68 +62,66 @@ def UnZipFile(zipFilePath, extractPath):
     cleanAndMakeDir(extractPath)
     for f in zipFile.namelist():
         zipFile.extract(f, extractPath)
+        
+def makeZipDif(newZip, oldZip, outPath):
+    print "makeZipDif(%s, %s, %s)" % (newZip, oldZip, outPath)
+    suffix = newZip[-3:]
+    resFolder = ""
+    if suffix == "apk":
+        resFolder = "assets/"
+    elif suffix == "ipa":
+        resFolder = "Payload/HGAME.app/"
+    
+    oldUnZipPath = "package/temp/old/"
+    UnZipFile(oldZip, oldUnZipPath)
+    newZipPath = "package/temp/new/"
+    UnZipFile(newZip, newZipPath)
+    bOnlyArr = []
+    diffArr = []
+    funnyArr = []
+    dirCmp(oldUnZipPath + resFolder + "res", newZipPath + resFolder + "res", bOnlyArr, diffArr, funnyArr)
+    dirCmp(oldUnZipPath + resFolder + "asset", newZipPath + resFolder + "asset", bOnlyArr, diffArr, funnyArr)
+    print "right-only files: %s" % (str(bOnlyArr))
+    print "diff files: %s" % (str(diffArr))
+    print "funny files: %s" % (str(funnyArr))    
+    updateZipFilePath = "package/temp/update.zip"
+    updateZipFile = zipfile.ZipFile(updateZipFilePath, 'w' ,zipfile.ZIP_DEFLATED)
+    for onlyF in bOnlyArr:
+        zipName = onlyF[len(newZipPath + resFolder):]
+        updateZipFile.write(onlyF, zipName)
+        print "update.zip add %s" % zipName
+    for diffF in diffArr:
+        zipName = diffF[len(newZipPath + resFolder):]
+        updateZipFile.write(diffF, zipName)
+        print "update.zip add %s" % zipName
+    updateZipFile.close()
+    
+    removeDir(oldUnZipPath)
+    removeDir(newZipPath)
+
+    copy(updateZipFilePath, outPath)
+    os.remove(updateZipFilePath)
 
 def testUnZip():
-    lastApkPath = "package/android/last.apk"
-    lastApkUnZipPath = "package/android/last"
-    UnZipFile(lastApkPath, lastApkUnZipPath)
-    newApkPath = "package/android/new.apk"
-    newApkUnZipPath = "package/android/new"
-    UnZipFile(newApkPath, newApkUnZipPath)
     
-    bOnlyArr = []
-    diffArr = []
-    funnyArr = []
-    dirCmp(lastApkUnZipPath + "/assets", newApkUnZipPath + "/assets", bOnlyArr, diffArr, funnyArr)
-    print "right-only files: %s" % (str(bOnlyArr))
-    print "diff files: %s" % (str(diffArr))
-    print "funny files: %s" % (str(funnyArr))
+    updateFolder = "../../test-Resources/package/update/"
+    platformArr = [ "android", "iOS" ]
+    platformPrefixArr = { "android": "ap", "iOS": "ip" }
     
-    updateApkZipPath = "package/android/update.zip"
-    updateApkZipFile = zipfile.ZipFile(updateApkZipPath, 'w' ,zipfile.ZIP_DEFLATED)
-    for onlyF in bOnlyArr:
-        zipName = onlyF[len(newApkUnZipPath + "/assets/"):]
-        updateApkZipFile.write(onlyF, zipName)
-        print "android update.zip add %s" % zipName
-    for diffF in diffArr:
-        zipName = diffF[len(newApkUnZipPath + "/assets/"):]
-        updateApkZipFile.write(diffF, zipName)
-        print "android update.zip add %s" % zipName
-    updateApkZipFile.close()
+    for platformName in platformArr:
+        packagePath = updateFolder + platformName + "/"
+        apkFileArr = open(packagePath + "package.txt").readlines()
+        lastFile = None
+        for apkFileName in apkFileArr:
+            apkFileName = str(apkFileName).strip()
+            updateFileName = apkFileName[:len(apkFileName) - 3]
+            updateFileName = updateFileName + "data"
+            updateFileName = updateFileName.replace("hgame.", platformPrefixArr[platformName])
+            if not lastFile == None:
+                makeZipDif(packagePath + apkFileName, packagePath + lastFile, packagePath + updateFileName)
+            lastFile = apkFileName
     
-    removeDir(lastApkUnZipPath)
-    removeDir(newApkUnZipPath)
-    
-    lastIPAPath = "package/iOS/last.ipa"
-    lastIPAUnZipPath = "package/iOS/last/"
-    UnZipFile(lastIPAPath, lastIPAUnZipPath)
-    newIPAPath = "package/iOS/new.ipa"
-    newIPAUnZipPath = "package/iOS/new/"
-    UnZipFile(newIPAPath, newIPAUnZipPath)
-    bOnlyArr = []
-    diffArr = []
-    funnyArr = []
-    dirCmp(lastIPAUnZipPath + "/Payload/HGAME.app/res", newIPAUnZipPath + "/Payload/HGAME.app/res", bOnlyArr, diffArr, funnyArr)
-    dirCmp(lastIPAUnZipPath + "/Payload/HGAME.app/asset", newIPAUnZipPath + "/Payload/HGAME.app/asset", bOnlyArr, diffArr, funnyArr)
-    print "right-only files: %s" % (str(bOnlyArr))
-    print "diff files: %s" % (str(diffArr))
-    print "funny files: %s" % (str(funnyArr))     
-    updateIPAZipPath = "package/iOS/update.zip"
-    updateIPAZipFile = zipfile.ZipFile(updateIPAZipPath, 'w' ,zipfile.ZIP_DEFLATED)
-    for onlyF in bOnlyArr:
-        zipName = onlyF[len(newIPAUnZipPath + "/Payload/HGAME.app/"):]
-        updateIPAZipFile.write(onlyF, zipName)
-        print "iOS update.zip add %s" % zipName
-    for diffF in diffArr:
-        zipName = diffF[len(newIPAUnZipPath + "/Payload/HGAME.app/"):]
-        updateIPAZipFile.write(diffF, zipName)
-        print "iOS update.zip add %s" % zipName
-    updateIPAZipFile.close()
-     
-    removeDir(lastIPAUnZipPath)
-    removeDir(newIPAUnZipPath)
 
 if __name__ == '__main__':
-#     testDirCmp()
     testUnZip()
